+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.2 (upcoming)
 
  - tweak layout
+ - add simple theme
 
 
 ## v0.0.1